Added check for pre-existing installation.

+++ setup.sh
@@ -15,26 +15,52 @@ init_vars() {
 
 # Test.
 test() {
-    printf '\nIP:'
+    printf '\nUser provided IP: '
     echo $ip_address
-    printf '\nGH: '
+    printf '\nGH Download URL: '
     echo $down_url
 }
 
 # Commands sent over SSH stdin as a heredoc.
 remote_install() {
     ssh root@$ip_address -oHostKeyAlgorithms=+ssh-rsa << ENDSSH
+
+    # Connection check.
+    if ping -c 1 1.1.1.1 &> /dev/null
+        then
+            echo "Device is connected to the internet."
+        else
+            echo "Device is not connected to the internet."
+            exit 0
+    fi
+
+    # Check to see if blue-merle is already installed.
+    if opkg list | grep blue-merle &> /dev/null
+        then
+            printf "\n\nAlready installed!\n"
+            exit 0
+        else
+            printf "\n\nStarting installation.\n"
+    fi
+
+    # Download and install.
+    echo "Downloading blue-merle."
     curl -L $down_url -o /tmp/blue-merle.ipk
     opkg update
     opkg install /tmp/blue-merle.ipk
-    reboot
-ENDSSH
-}
 
-# Post-install messages.
-post_install() {
-    printf '\n\nInstall complete, device will now reboot!\n'
-    printf '\nAfter device boots:\nFlip side-switch to the up position (towards recessed dot) and follow on-device MCU prompts.\n\n'
+    # Error Check to see if blue-merle is installed.
+    if opkg list | grep blue-merle &> /dev/null
+        then
+            printf "\n\nInstall complete, device will now reboot!\n"
+            printf '\nAfter device boots:\nFlip side-switch to the up position (towards recessed dot) and follow on-device MCU prompts.\n\n'
+            sleep 1
+            reboot
+        else
+            printf "\n\nInstall Failed!\n"
+            exit 0
+    fi
+ENDSSH
 }
 
 # Main.
@@ -42,4 +68,3 @@ pre_install
 init_vars
 #test
 remote_install
-post_install
